Edit Profile Password
The password empty field is getting dropped and the full field is
getting salted. So the user can now log back in after editing their
profile and submitting a blank password field. Also, they can change
their password and it will get salted. Moving on to the unique email
check.