p_delete does delete the post but ...
But it generates a page of errors. If I remove the p from p_delete, I
see the delete confirmation page but the delete button doesn't work.

I had broken something when mokeying around with the delete post method
so I went back two commits and downloaded a folder that wasn't broken.
Gitapp didn't like that very much so I need to create a new repository.
I think I had to do this before … and I do have a back up of this file
outside of sites folder.